+++ CHANGES.md
@@ -1,5 +1,6 @@
 Changes coming in 0.3.0-alpha6 (or beta1)
 
+* query as-arrays? now allows you to leverage lazy result fetching [JDBC-72](http://dev.clojure.org/jira/browse/JDBC-72).
 * "h2" is recognized as a protocol shorthand for org.h2.Driver
 * Tests no longer use :1 literal [JDBC-71](http://dev.clojure.org/jira/browse/JDBC-71).